Moved receiver Main to a dedicated package
